Initial version
It downloads missing images and runs sha256 on them.  By using `make
clean`, the sha256s are cleared but not the bulk data. By using "-sj
N" you can get concurrency.  Because it's a Makefile, if you have half
the images, but not the other half, it offers a form of "resume,"
though no attempt is made to recover partial files.  Likewise, to
re-download and re-calculate the sha256 for a given image, delete it
and run `make`, it will only do the download for the missing image.